Move alarm list out of QAM into alarm list file
